Refactor visual element url transformer
